support ALL services keyword

+++ system/scripts/update/finish_engines_system_update.sh
@@ -19,18 +19,28 @@ echo Applying update $update_id &>> $system_updates_dir/$update_id/update_log
  if test -f  $system_updates_dir/$update_id/services
   then
   	services=`cat  $system_updates_dir/$update_id/services`
+  	echo $services |grep ALL >/dev/null
+  	if test $? -eq 0
+  	 then
+  	   services=`ls /opt/engines/run/services`
+  	 fi 
+  	 
   		for service in $services
   		 do  		 
   		 if test -f /opt/engines/run/services/$service/running.yaml
   		  then
+  		    if  ! test `/opt/engines/bin/engines service $service state` = nocontainer
+  		     then  		      
   		 	/opt/engines/bin/engines service $service stop &>> $system_updates_dir/$update_id/update_log 		 	 
   		 	image=`grep image /opt/engines/run/services/$service/running.yaml | cut -f2 -d" "`
   			docker pull $image
+  			
   			/opt/engines/bin/engines service $service wait_for stop 60
   			rm /opt/engines/run/services/$service/running.yaml*
   		 	/opt/engines/bin/engines service $service recreate &>>  $system_updates_dir/$update_id/update_log
   		 	/opt/engines/bin/engines service $service wait_for start 30
-			/opt/engines/bin/engines service $service wait_for_startup 60			
+			/opt/engines/bin/engines service $service wait_for_startup 60
+			fi			
 		  fi
   		 done
   fi
